Add files via upload
The app is a web-based application that allows users to add and store their secrets in a safe and secure manner. It has been created using a combination of technologies such as EJS, Express.js, MongoDB, and Passport.js. The app has been developed by referencing and taking help from the renowned instructor, Angela Yu.

Using the app, users can create a new account, log in to their existing account, and add their secrets. The app uses Passport.js for user authentication, which ensures that only authorized users can access their secrets. The secrets are stored in MongoDB, which ensures that they are stored securely and can be accessed by the user at any time.

The app has a clean and user-friendly interface designed using EJS, which allows the user to easily navigate and use the app. Users can add, edit and delete their secrets, and can also view all the secrets added by them.

Overall, the app provides a secure and reliable way for users to store and manage their secrets, with the help of advanced technologies and best practices implemented by the experienced instructor Angela Yu.